simplify the Jaeger integration test
